refactor: move a utility out of openedx_webhooks.auth
